Backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
reportlab
